+++ DentaCare/src/java/controller/StaffViewBooking.java
@@ -63,7 +63,7 @@ public class StaffViewBooking extends HttpServlet {
                 java.sql.Date sqlNextDate = new java.sql.Date(nextDate.getTime());
                 List<BookingDTO> listBookingNextDate = daoBooking.getAllBookingClinic(staff.getClinicID(), sqlNextDate);
                  //Show revenue
-                Double revenue = daoStaffAccount.getRevenue(now);
+                Long revenue = daoStaffAccount.getRevenue(now);
                 //Set Attribute
                 request.setAttribute("style", "none");
                 request.setAttribute("dailyRevenue", revenue);
@@ -89,7 +89,7 @@ public class StaffViewBooking extends HttpServlet {
                 java.sql.Date sqlNextDate = new java.sql.Date(nextDate.getTime());
                 List<BookingDTO> listBookingNextDate = daoBooking.getAllBookingClinic(staff.getClinicID(), sqlNextDate);
                 //Show revenue
-                Double revenue = daoStaffAccount.getRevenue(now);
+                Long revenue = daoStaffAccount.getRevenue(now);
 
                 //Set Attribute
                 request.setAttribute("style", "none");
@@ -164,7 +164,7 @@ public class StaffViewBooking extends HttpServlet {
                 java.sql.Date sqlNextDate = new java.sql.Date(nextDate.getTime());
                 List<BookingDTO> listBookingNextDate = daoBooking.getAllBookingClinic(staff.getClinicID(), sqlNextDate);
                 //Show revenue
-                Double revenue = daoStaffAccount.getRevenue(now);
+                Long revenue = daoStaffAccount.getRevenue(now);
                 //Set Attribute
                 request.setAttribute("style", "none");
                 request.setAttribute("dailyRevenue", revenue);
